Compact arena: query API, SIMD name-id scan, construction + allocation fixes (#19)
* Compact arena: intern names at construction, lazy leaf wrappers, SIMD name-id queries

- Name interning moved to construction into a ushort NameId column; drop the
  StringOrMemory Names column and the freeze re-walk (frozen doc reads ids from
  columns directly). Construction-only columns released at freeze.
- CompactSourceLocation Line/Column widened ushort -> int (no overflow foot-gun).
- Leaf (text/comment/PI) wrappers no longer allocated eagerly; Node(handle)
  materializes lazily. ~32% less parse allocation on text-heavy docs.
- CompactDocument.IndexOfName: vectorized name-id scan (ushort column reinterpreted
  as char, SIMD IndexOf) + scalar packed fallback.
- CompactQuery: predicate-pushdown-lite fluent helpers (tag pushed into SIMD scan,
  class/attribute filters applied to candidates only), zero-alloc struct enumerator.
- Benchmarks: FatDocument source-tracking tier (explicit config, og parser commented),
  FullFlow POCO extraction (scalar + SIMD arena), Selectivity SIMD find microbench.


* Query API for the compact arena + RODOM element-only fast path

- Node cursor (readonly struct): Is/Attr/HasClass/Children/Text over a (doc, handle) pair;
  string, ReadOnlySpan<char>, and pre-resolved-id overloads.
- CompactQuery: Descendants() familiar scan + Elements(tag).WithClass/WithAttribute
  pushdown-lite (tag pushed into the SIMD name-id scan, filters applied to candidates only),
  zero-alloc struct enumerators.
- Split name resolution: ResolveNameId (pure, O(1) for known names) for the query hot path
  vs FindNameId (presence-checked existence, O(nodes)). The query layer had been calling the
  presence scanner per node, making descendant queries accidentally O(nodes^2).
- CompactDocument.KindAt/NameIdAt single-column reads so Node.Is avoids reconstructing a
  whole node.
- Text sinks: WriteText<TSink>(ISpanSink) plus TextWriter / IBufferWriter<char> / Span<char>
  / length paths, streaming value spans with no intermediate string.
- QueryHelpers.CountTagClass: additive concrete-typed, element-only RODOM walk (existing
  helpers untouched).
- Benchmarks: FullFlow rewired to the cursor API; Query and Selectivity query-only benches.


---------

+++ AngleSharp.ReadOnlyDom.CompactPrototype/CompactNode.cs
@@ -71,7 +71,7 @@ public readonly struct CompactAttribute
     public int ValueLength { get; }
 }
 
-public readonly record struct CompactSourceLocation(int Index, ushort Line, ushort Column);
+public readonly record struct CompactSourceLocation(int Index, int Line, int Column);
 
 [Flags]
 public enum CompactMetadataOptions
